LPD-56365 Split to avoid having big test files
