+++ modules/test/playwright/tests/site-cms-site-initializer/structure-builder/frontendValidations.spec.ts
@@ -1,0 +1,224 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+import {expect, mergeTests} from '@playwright/test';
+
+import {featureFlagsTest} from '../../../fixtures/featureFlagsTest';
+import {loginTest} from '../../../fixtures/loginTest';
+import {pageEditorPagesTest} from '../../../fixtures/pageEditorPagesTest';
+import {clickAndExpectToBeHidden} from '../../../utils/clickAndExpectToBeHidden';
+import {clickAndExpectToBeVisible} from '../../../utils/clickAndExpectToBeVisible';
+import {getRandomInt} from '../../../utils/getRandomInt';
+import {waitForAlert} from '../../../utils/waitForAlert';
+import {cmsPagesTest} from '../main/fixtures/cmsPagesTest';
+import {structureBuilderPagesTest} from './fixtures/structureBuilderPagesTest';
+
+const test = mergeTests(
+	cmsPagesTest,
+	featureFlagsTest({
+		'LPD-17564': {enabled: true},
+		'LPS-179669': {enabled: true},
+	}),
+	loginTest(),
+	pageEditorPagesTest,
+	structureBuilderPagesTest
+);
+
+let structureIds = [];
+
+test.beforeEach(() => {
+	structureIds = [];
+});
+
+test.afterEach(async ({structureBuilderPage}) => {
+	for (const id of structureIds) {
+		await structureBuilderPage.deleteStructure(Number(id));
+	}
+});
+
+test(
+	'Validations when saving the structure',
+	{tag: '@LPD-36752'},
+	async ({page, picklistBuilderPage, structureBuilderPage}) => {
+
+		// Add a picklist
+
+		const picklist = await picklistBuilderPage.createPicklist();
+
+		// Go to the Structure Builder
+
+		await structureBuilderPage.goToCreateStructure();
+
+		// Add a Text field
+
+		await structureBuilderPage.addField('Text');
+
+		// Try to save and check we can't publish without spaces
+
+		await expect(async () => {
+			await structureBuilderPage.saveButton.click();
+
+			await expect(
+				page.getByText('Spaces must be selected')
+			).toBeAttached({
+				timeout: 500,
+			});
+		}).toPass();
+
+		await structureBuilderPage.enableForAllSpaces();
+
+		// Set label and empty name
+
+		const label = `Structure${getRandomInt()}`;
+
+		await structureBuilderPage.changeStructureSettings({
+			label,
+			name: '',
+		});
+
+		await expect(page.getByText('This field is required')).toBeVisible();
+
+		// Add a Single Select field and select it
+
+		await structureBuilderPage.addField('Single Select');
+
+		await structureBuilderPage.selectFields([{label: 'Single Select'}]);
+
+		// Put empty name
+
+		await structureBuilderPage.changeFieldSettings({name: ''});
+
+		// Try to save and check it redirects to structure view
+
+		await clickAndExpectToBeVisible({
+			target: page.getByText('Structure Name'),
+			trigger: structureBuilderPage.saveButton,
+		});
+
+		// Fill name
+
+		await structureBuilderPage.changeStructureSettings({name: label});
+
+		// Now try to save and check it redirects to field view
+
+		await clickAndExpectToBeVisible({
+			target: page.locator('.breadcrumb-link', {
+				hasText: 'Single Select',
+			}),
+			trigger: structureBuilderPage.saveButton,
+		});
+
+		// Fill name, select picklist and save again
+
+		await structureBuilderPage.changeFieldSettings({name: 'name'});
+
+		await structureBuilderPage.changeFieldSettings({
+			picklist: picklist.name,
+		});
+
+		// Check picklist setting is saved
+
+		await structureBuilderPage.selectFields([{label: 'Text'}]);
+
+		await structureBuilderPage.selectFields([{label: 'Single Select'}]);
+
+		await expect(page.getByText(picklist.name)).toBeVisible();
+
+		// Save
+
+		const {id} = await structureBuilderPage.saveStructure();
+
+		structureIds.push(id);
+
+		// Publish structure
+
+		await structureBuilderPage.publishStructure();
+
+		// Delete one field and check warning modal is show when publishing
+
+		await structureBuilderPage.deleteFields([{label: 'Text'}]);
+
+		await clickAndExpectToBeVisible({
+			target: page.getByText(
+				'You removed one or more fields from the structure'
+			),
+			trigger: structureBuilderPage.publishButton,
+		});
+
+		await clickAndExpectToBeHidden({
+			target: page.getByText(
+				'You removed one or more fields from the structure'
+			),
+			trigger: page.locator('.btn-danger'),
+		});
+
+		await waitForAlert(page, 'published successfully', {
+			timeout: 5000,
+		});
+
+		// Check the warning does not appear anymore
+
+		await structureBuilderPage.publishStructure();
+
+		// Delete picklist
+
+		await picklistBuilderPage.deletePicklist(picklist.id);
+	}
+);
+
+test(
+	'Validations in the picklist picker',
+	{tag: '@LPD-51647'},
+	async ({page, picklistBuilderPage, structureBuilderPage}) => {
+
+		// Add a picklist
+
+		const picklist = await picklistBuilderPage.createPicklist();
+
+		// Go to the Structure Builder
+
+		await structureBuilderPage.goToCreateStructure();
+
+		// Add a Single Select field and check for blur error
+
+		await structureBuilderPage.addField('Single Select');
+
+		await structureBuilderPage.selectFields([{label: 'Single Select'}]);
+
+		const picklistPicker = page.getByLabel('Picklist');
+
+		const errorMessage = page.getByText('This field is required.');
+
+		await expect(errorMessage).not.toBeAttached();
+
+		await picklistPicker.press('Tab');
+
+		await expect(errorMessage).toBeAttached();
+
+		await structureBuilderPage.changeFieldSettings({
+			picklist: picklist.name,
+		});
+
+		await expect(errorMessage).not.toBeAttached();
+
+		// Add a Multiselect field and check for outside click error
+
+		await structureBuilderPage.addField('Multiselect');
+
+		await structureBuilderPage.selectFields([{label: 'Multiselect'}]);
+
+		await expect(errorMessage).not.toBeAttached();
+
+		await picklistPicker.click();
+
+		await page.locator('body').click();
+
+		await expect(errorMessage).toBeAttached();
+
+		// Delete picklist
+
+		await picklistBuilderPage.deletePicklist(picklist.id);
+	}
+);
