+++ modules/test/playwright/tests/site-cms-site-initializer/structure-builder/repeatableGroups.spec.ts
@@ -1,0 +1,277 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+import {expect, mergeTests} from '@playwright/test';
+
+import {featureFlagsTest} from '../../../fixtures/featureFlagsTest';
+import {loginTest} from '../../../fixtures/loginTest';
+import {pageEditorPagesTest} from '../../../fixtures/pageEditorPagesTest';
+import {clickAndExpectToBeHidden} from '../../../utils/clickAndExpectToBeHidden';
+import {clickAndExpectToBeVisible} from '../../../utils/clickAndExpectToBeVisible';
+import {getRandomInt} from '../../../utils/getRandomInt';
+import getRandomString from '../../../utils/getRandomString';
+import {cmsPagesTest} from '../main/fixtures/cmsPagesTest';
+import {structureBuilderPagesTest} from './fixtures/structureBuilderPagesTest';
+
+const test = mergeTests(
+	cmsPagesTest,
+	featureFlagsTest({
+		'LPD-17564': {enabled: true},
+		'LPS-179669': {enabled: true},
+	}),
+	loginTest(),
+	pageEditorPagesTest,
+	structureBuilderPagesTest
+);
+
+let structureIds = [];
+
+test.beforeEach(() => {
+	structureIds = [];
+});
+
+test.afterEach(async ({structureBuilderPage}) => {
+	for (const id of structureIds) {
+		await structureBuilderPage.deleteStructure(Number(id));
+	}
+});
+
+test(
+	'Groups can be created, persisted and ungrouped',
+	{
+		tag: '@LPD-50378',
+	},
+	async ({page, structureBuilderPage}) => {
+
+		// Create structure
+
+		const erc = await structureBuilderPage.createStructureFromData({
+			label: getRandomString(),
+			name: `StructureName${getRandomInt()}`,
+			page: structureBuilderPage,
+			publish: false,
+			structureIds,
+		});
+
+		// Add fields
+
+		await structureBuilderPage.addField('Text');
+		await structureBuilderPage.addField('Date');
+		await structureBuilderPage.addField('Decimal');
+
+		// Create repeatable group with two of them
+
+		await structureBuilderPage.createRepeatableGroup({
+			fields: [{label: 'Text'}, {label: 'Date'}],
+			label: 'Repeatable Group 1',
+		});
+
+		// Check recently added group is expanded by default
+
+		await expect(
+			page.locator('.treeview-link', {hasText: 'Text'})
+		).toBeVisible();
+
+		await expect(
+			page.locator('.treeview-link', {hasText: 'Date'})
+		).toBeVisible();
+
+		// Create another group inside the first one
+
+		await structureBuilderPage.createRepeatableGroup({
+			fields: [{label: 'Date'}],
+			label: 'Repeatable Group 2',
+		});
+
+		// Assert correct label style
+
+		await expect(
+			page.locator('.label-success', {hasText: 'Repeatable Group'})
+		).toBeVisible();
+
+		// Check groups are persisted
+
+		await structureBuilderPage.publishStructure();
+
+		await structureBuilderPage.editStructure(erc);
+
+		await expect(
+			page.locator('.treeview-link', {hasText: 'Repeatable Group 1'})
+		).toBeVisible();
+
+		await structureBuilderPage.expandField({
+			label: 'Repeatable Group 1',
+		});
+
+		await expect(
+			page.locator('.treeview-link', {hasText: 'Repeatable Group 2'})
+		).toBeVisible();
+
+		// Add a new group and ungroup it
+
+		await structureBuilderPage.addField('Boolean');
+
+		await structureBuilderPage.createRepeatableGroup({
+			fields: [{label: 'Boolean'}],
+			label: 'Repeatable Group 3',
+		});
+
+		await structureBuilderPage.clickFieldAction(
+			{label: 'Repeatable Group 3'},
+			'Ungroup'
+		);
+
+		await expect(
+			page.locator('.treeview-link', {hasText: 'Repeatable Group 2'})
+		).toBeVisible();
+
+		await expect(
+			page.locator('.treeview-link', {hasText: 'Boolean'})
+		).toBeVisible();
+	}
+);
+
+test(
+	'Check restrictions for group creation',
+	{
+		tag: '@LPD-50378',
+	},
+	async ({page, structureBuilderPage}) => {
+
+		// Create structure
+
+		const erc = await structureBuilderPage.createStructureFromData({
+			label: getRandomString(),
+			name: `StructureName${getRandomInt()}`,
+			page: structureBuilderPage,
+			publish: false,
+			structureIds,
+		});
+
+		// Check a group can't be created if there's only one field
+
+		await structureBuilderPage.selectFields([{label: 'Title'}]);
+
+		await structureBuilderPage.clickFieldAction(
+			{label: 'Title'},
+			'Create Repeatable Group'
+		);
+
+		await clickAndExpectToBeVisible({
+			target: page.getByText(
+				'The repeatable group cannot be created because at least one field is required.'
+			),
+			trigger: page.getByRole('menuitem', {
+				name: 'Create Repeatable Group',
+			}),
+		});
+
+		await clickAndExpectToBeHidden({
+			target: page.getByText(
+				'The repeatable group cannot be created because at least one field is required.'
+			),
+			trigger: page.locator('.modal-footer').getByText('Done'),
+		});
+
+		// Add fields
+
+		await structureBuilderPage.addField('Text');
+		await structureBuilderPage.addField('Date');
+		await structureBuilderPage.addField('Decimal');
+
+		// Create repeatable group with two of them
+
+		await structureBuilderPage.createRepeatableGroup({
+			fields: [{label: 'Text'}, {label: 'Date'}],
+			label: 'Repeatable Group 1',
+		});
+
+		// Check a group can't be created with fields that have different parent
+
+		await structureBuilderPage.selectFields([
+			{label: 'Date'},
+			{label: 'Decimal'},
+		]);
+
+		await clickAndExpectToBeVisible({
+			target: page.getByRole('menuitem', {
+				name: 'Create Repeatable Group',
+			}),
+			trigger: page.getByLabel('Selection Options'),
+		});
+
+		await clickAndExpectToBeVisible({
+			target: page.getByText(
+				'A repeatable group requires all selected items to be at the same hierarchy level. Adjust your selection and try again.'
+			),
+			trigger: page.getByRole('menuitem', {
+				name: 'Create Repeatable Group',
+			}),
+		});
+
+		await clickAndExpectToBeHidden({
+			target: page.getByText(
+				'A repeatable group requires all selected items to be at the same hierarchy level. Adjust your selection and try again.'
+			),
+			trigger: page.locator('.modal-footer').getByText('Done'),
+		});
+
+		// Check a group can't be created with published fields
+
+		await structureBuilderPage.publishStructure();
+
+		await structureBuilderPage.selectFields([
+			{label: 'Title'},
+			{label: 'Decimal'},
+		]);
+
+		await clickAndExpectToBeVisible({
+			target: page.getByRole('menuitem', {
+				name: 'Create Repeatable Group',
+			}),
+			trigger: page.getByLabel('Selection Options'),
+		});
+
+		await clickAndExpectToBeVisible({
+			target: page.getByText(
+				'The repeatable group cannot be created because one or more fields of the selection are already published.'
+			),
+			trigger: page.getByRole('menuitem', {
+				name: 'Create Repeatable Group',
+			}),
+		});
+
+		await clickAndExpectToBeHidden({
+			target: page.getByText(
+				'The repeatable group cannot be created because one or more fields of the selection are already published.'
+			),
+			trigger: page.locator('.modal-footer').getByText('Done'),
+		});
+
+		// Check we can't ungroup the published group
+
+		await structureBuilderPage.publishStructure();
+
+		await structureBuilderPage.editStructure(erc);
+
+		await structureBuilderPage.clickFieldAction(
+			{label: 'Repeatable Group 1'},
+			'Ungroup'
+		);
+
+		await page
+			.getByText(
+				'The ungroup action cannot be done because this repeatable group is already published.'
+			)
+			.waitFor();
+
+		await clickAndExpectToBeHidden({
+			target: page.getByText(
+				'The ungroup action cannot be done because this repeatable group is already published.'
+			),
+			trigger: page.locator('.modal-footer').getByText('Done'),
+		});
+	}
+);
